MatchingGodPageにMVVMを適用した
